Add files via upload
Credit scoring is a technique used by lenders and creditors to determine the risk of lending money to an individual or business. Credit scores are typically calculated based on the user's credit history, including their history of paying bills and loans.

A credit scoring system uses this information to create a unique credit score for each individual or business. The credit score serves as a prediction of how likely it is that a particular individual or business will repay their debts on time.

There are many types of credit scoring models, each with its own unique approach to determining creditworthiness. The most commonly used credit scoring models are FICO (Fair Isaac and Company's Credit System), VantageScore, and Experian Credit Score.